requirements.txt
gradio
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-generativeai
fastapi
uvicorn
python-dotenv
